backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary>=2.9.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
langchain>=0.3.7
langchain-google-genai>=2.0.5
langsmith>=0.1.147
pydantic[email]==2.9.2
pydantic-settings==2.6.1
email-validator>=2.0.0
python-multipart==0.0.12
langgraph==0.2.61
openpyxl==3.1.5
google-generativeai>=0.8.0
pytest>=8.0.0
httpx>=0.27.0
